Extract Desktop from InitiativesShow
